backend/requirements.txt
elasticsearch>=8.15,<9
fastapi
uvicorn[standard]
requests~=2.26.0
sentence-transformers>=3.0.1,<4
Pillow>=10.3.0
transformers>=4.44,<5
tokenizers>=0.20.0
PyYAML
python-dotenv>=1.0.1